# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/request-otp:
    post:
      operationId: requestOtp
      tags: [auth]
      summary: Request a phone OTP
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RequestOtpInput"
      responses:
        "200":
          description: OTP requested
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OtpRequestResult"
  /auth/verify-otp:
    post:
      operationId: verifyOtp
      tags: [auth]
      summary: Verify a phone OTP
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyOtpInput"
      responses:
        "200":
          description: OTP verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSession"
  /me:
    get:
      operationId: getCurrentUser
      tags: [users]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
    patch:
      operationId: updateCurrentUser
      tags: [users]
      summary: Update current authenticated user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileInput"
      responses:
        "200":
          description: Updated user profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
  /presence/heartbeat:
    post:
      operationId: updatePresenceHeartbeat
      tags: [users]
      summary: Update current user presence heartbeat
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PresenceHeartbeatInput"
      responses:
        "200":
          description: Presence updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PresenceSnapshot"
...
  /media/audio/upload-target:
    post:
      operationId: createAudioUploadTarget
      tags: [media]
      summary: Create an upload target for an audio note
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAudioUploadTargetInput"
      responses:
        "200":
          description: Audio upload target
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AudioUploadTarget"
  /media/upload-target:
    post:
      operationId: createMediaUploadTarget
      tags: [media]
      summary: Create an upload target for avatars, chat media or stories
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMediaUploadTargetInput"
      responses:
        "200":
          description: Generic media upload target
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MediaUploadTarget"
  /media/resolve:
    get:
      operationId: resolveMediaUrl
      tags: [media]
      summary: Resolve a media key to a public URL
      parameters:
        - name: key
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Resolved media URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResolvedMedia"
  /stories:
    get:
      operationId: listStories
      tags: [stories]
      summary: List active stories
      responses:
        "200":
          description: Story list
          content:
            application/json:
              schema:
                type: object
                properties:
                  stories:
                    type: array
                    items:
                      $ref: "#/components/schemas/Story"
                required:
                  - stories
    post:
      operationId: createStory
      tags: [stories]
      summary: Create a story
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStoryInput"
      responses:
        "200":
          description: Created story
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Story"
  /stories/{storyId}/views:
    post:
      operationId: addStoryView
      tags: [stories]
      summary: Mark a story as viewed
      parameters:
        - $ref: "#/components/parameters/StoryId"
      responses:
        "200":
          description: Story with updated viewers
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Story"
components:
  parameters:
...
    StoryId:
      name: storyId
      in: path
      required: true
      schema:
        type: string
  schemas:
...
    UserSettings:
      type: object
      properties:
        lastSeenVisibility:
          type: string
          enum: [everyone, contacts, nobody]
        readReceiptsEnabled:
          type: boolean
        notificationsEnabled:
          type: boolean
        notificationSoundEnabled:
          type: boolean
        vibrationEnabled:
          type: boolean
        autoDownloadMedia:
          type: boolean
        lowDataMode:
          type: boolean
        chatFontScale:
          type: string
          enum: [small, medium, large]
      required:
        - lastSeenVisibility
        - readReceiptsEnabled
        - notificationsEnabled
        - notificationSoundEnabled
        - vibrationEnabled
        - autoDownloadMedia
        - lowDataMode
        - chatFontScale
    PresenceSnapshot:
      type: object
      properties:
        isOnline:
          type: boolean
        lastSeenAt:
          type: string
          format: date-time
          nullable: true
      required:
        - isOnline
        - lastSeenAt
    UserProfile:
      type: object
      properties:
        id:
          type: string
        phone:
          type: string
        countryCode:
          type: string
        name:
          type: string
        avatarUrl:
          type: string
          nullable: true
        bio:
          type: string
        statusText:
          type: string
        initials:
          type: string
        color:
          type: string
        isOnboarded:
          type: boolean
        settings:
          $ref: "#/components/schemas/UserSettings"
        presence:
          $ref: "#/components/schemas/PresenceSnapshot"
      required:
        - id
        - phone
        - countryCode
        - name
        - avatarUrl
        - bio
        - statusText
        - initials
        - color
        - isOnboarded
        - settings
        - presence
    RequestOtpInput:
      type: object
      properties:
        phone:
          type: string
        countryCode:
          type: string
      required:
        - phone
        - countryCode
    OtpRequestResult:
      type: object
      properties:
        requestId:
          type: string
        expiresAt:
          type: string
          format: date-time
        demoCode:
          type: string
          nullable: true
      required:
        - requestId
        - expiresAt
        - demoCode
    VerifyOtpInput:
      type: object
      properties:
        requestId:
          type: string
        phone:
          type: string
        code:
          type: string
      required:
        - requestId
        - phone
        - code
    AuthSession:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/UserProfile"
      required:
        - token
        - user
    UpdateProfileInput:
      type: object
      properties:
        name:
          type: string
        avatarUrl:
          type: string
          nullable: true
        bio:
          type: string
        statusText:
          type: string
        notificationsEnabled:
          type: boolean
        notificationSoundEnabled:
          type: boolean
        vibrationEnabled:
          type: boolean
        autoDownloadMedia:
          type: boolean
        lowDataMode:
          type: boolean
        readReceiptsEnabled:
          type: boolean
        lastSeenVisibility:
          type: string
          enum: [everyone, contacts, nobody]
        chatFontScale:
          type: string
          enum: [small, medium, large]
      additionalProperties: false
    PresenceHeartbeatInput:
      type: object
      properties:
        isOnline:
          type: boolean
      required:
        - isOnline
...
    MediaCategory:
      type: string
      enum: [audio, avatar, chat-image, chat-video, story-image, story-video]
    CreateMediaUploadTargetInput:
      type: object
      properties:
        category:
          $ref: "#/components/schemas/MediaCategory"
        mimeType:
          type: string
        conversationId:
          type: string
          nullable: true
      required:
        - category
        - mimeType
    MediaUploadTarget:
      type: object
      properties:
        key:
          type: string
        uploadUrl:
          type: string
        publicUrl:
          type: string
        contentType:
          type: string
        category:
          $ref: "#/components/schemas/MediaCategory"
      required:
        - key
        - uploadUrl
        - publicUrl
        - contentType
        - category
    ResolvedMedia:
      type: object
      properties:
        key:
          type: string
        url:
          type: string
      required:
        - key
        - url
    CreateAudioUploadTargetInput:
      type: object
      properties:
        mimeType:
          type: string
      required:
        - mimeType
    AudioUploadTarget:
      type: object
      properties:
        key:
          type: string
        uploadUrl:
          type: string
        publicUrl:
          type: string
        contentType:
          type: string
      required:
        - key
        - uploadUrl
        - publicUrl
        - contentType
    Story:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        type:
          type: string
          enum: [text, image, video]
        content:
          type: string
        backgroundColor:
          type: string
        expiresAt:
          type: string
          format: date-time
        viewerIds:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - type
        - content
        - backgroundColor
        - expiresAt
        - viewerIds
        - createdAt
    CreateStoryInput:
      type: object
      properties:
        type:
          type: string
          enum: [text, image, video]
        content:
          type: string
        backgroundColor:
          type: string
      required:
        - type
        - content
...